add more icons
