+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lascaux-sketch",
-  "version": "0.1.2",
+  "version": "0.1.3",
   "license": "Zlib",
   "author": {
     "name": "Marcello Bastea-Forte",
