+++ src/lascaux-ui/IconUrls.ts
@@ -1,0 +1,9 @@
+export type IconsUrls = {
+  layerPlus: string;
+  play: string;
+  pause: string;
+  undo: string;
+  redo: string;
+  layer: string;
+  selectedLayer: string;
+};
